// frontend/src/app/api/model/syncChangeItem.ts
/**
 * Life Management 2.0 API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface SyncChangeItem { 
    entityType: string;
    id: string;
    deleted: boolean;
    updatedAt: string;
    /**
     * Same shape as `GET` by id for this entity type. Null for tombstones (deleted = true).
     */
    data?: object | null;
}